+++ src/layouts/InterfaceLayout/containers/SendCurrencyContainer/SendCurrencyContainer.vue
@@ -173,7 +173,7 @@ import CurrencyPicker from '../../components/CurrencyPicker';
 import InterfaceBottomText from '@/components/InterfaceBottomText';
 import Blockie from '@/components/Blockie';
 import EthTx from 'ethereumjs-tx';
-import { Misc } from '@/helpers';
+import { Misc, ErrorHandler } from '@/helpers';
 import BigNumber from 'bignumber.js';
 import ethUnit from 'ethjs-unit';
 import utils from 'web3-utils';
@@ -241,7 +241,7 @@ export default {
       return (
         this.isValidAmount &&
         this.isValidAddress &&
-        new BigNumber(this.gasLimit).gte(0) &&
+        (new BigNumber(this.gasLimit).gte(0) || this.gasLimit == -1) &&
         Misc.validateHexString(this.data)
       );
     },
@@ -331,7 +331,15 @@ export default {
         ),
         data: this.txData
       };
-      this.gasLimit = await this.web3.eth.estimateGas(params);
+      this.web3.eth
+        .estimateGas(params)
+        .then(gasLimit => {
+          this.gasLimit = gasLimit;
+        })
+        .catch(err => {
+          this.gasLimit = -1;
+          ErrorHandler(err, true);
+        });
     },
     async submitTransaction() {
       window.scrollTo(0, 0);
@@ -351,7 +359,9 @@ export default {
         });
         const json = _tx.toJSON(true);
         json.from = coinbase;
-        this.web3.eth.sendTransaction(json);
+        this.web3.eth.sendTransaction(json).catch(err => {
+          ErrorHandler(err, true);
+        });
       } catch (e) {
         throw e;
       }