+++ src/networks/types/TOMO.js
@@ -8,10 +8,8 @@ export default {
   homePage: 'https://tomochain.com/',
   blockExplorerTX: 'https://explorer.tomocoin.io/#/tx/[[txHash]]',
   blockExplorerAddr: 'https://explorer.tomocoin.io/#/address/[[address]]',
-  chainID: 40686,
+  chainID: 88,
   tokens: tokens,
   contracts: contracts,
-  ensResolver: '',
-  ensAbi: '',
   icon: tomo
 };